Add files via upload
A new floor was added to the science building at our university, and while waiting for the elevator, a senior mentioned that the waiting time had not only increased but had increased non-linearly. To verify this, I attempted to simulate the problem. Initially, my approach was incorrect, but the senior suggested using a time-step method, which proved to be effective. Through this approach, I was able to confirm the non-linear nature of the increased waiting time.